third_party/blink/web_tests/svg/dynamic-updates/SVGLinearGradientElement-dom-gradientTransform-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'gradientTransform' attribute of the SVGLinearGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS linearGradientElement.getAttribute('gradientTransform') is "matrix(1,0,0,1,50,0)"
PASS linearGradientElement.getAttribute('gradientTransform') is "matrix(1,0,0,1,-100,0)"
PASS successfullyParsed is true

TEST COMPLETE
